% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{get_orders}
\alias{get_orders}
\title{Get orders}
\usage{
get_orders(GH = NULL, pageNum = 1, pageSize = 20, ...)
}
\arguments{
\item{...}{...}
}
\value{
Dataframe of orders
}
\description{
Get all past orders.
}
\details{
Builds a special URL to look at all of your orders referred to by:
}
\examples{
\dontrun{

get_orders(gh_auth(), 90, 1)


}
}
